Reset Functionality
The in game messages only show up when certain parameters are met. If an artefact has been collected it will update an array of bools in game manager. Collectible artefacts need an ID now when attaching the pick up script. Chat to me for setting things up or if you have problems.